add test to test_utils
